Moved ApiMocking class
