+++ testutils/build.gradle
@@ -20,6 +20,7 @@ dependencies {
     implementation project(path: ':core')
     implementation dependency.aws.mobileclient
     implementation dependency.junit
+    implementation dependency.mockito
     implementation dependency.androidx.test.core
     implementation dependency.rxjava
